更改 c++ java makefile 文件以后暂时能编译通过，必须有bug 待测试

+++ HelloDemo/frameworks/cocos2d-x/cocos/audio/winrt/AudioSourceReader.h
@@ -2,6 +2,7 @@
 * cocos2d-x   http://www.cocos2d-x.org
 *
 * Copyright (c) 2010-2011 - cocos2d-x community
+* Copyright (c) 2017-2018 Xiamen Yaji Software Co., Ltd.
 *
 * Portions Copyright (c) Microsoft Open Technologies, Inc.
 * All Rights Reserved
@@ -30,16 +31,16 @@
 #include <mfreadwrite.h>
 #include <queue>
 #include <mutex>
-#include "MediaStreamer.h"
+#include "audio/winrt/MediaStreamer.h"
 #include "ogg/ogg.h"
 #include "vorbis/vorbisfile.h"
 
 NS_CC_BEGIN
 namespace experimental{
 
-const UINT PCMDATA_CACHEMAXSIZE = 2621440;
-const UINT QUEUEBUFFER_NUM = 4;
-const UINT CHUNK_SIZE_MAX = PCMDATA_CACHEMAXSIZE / QUEUEBUFFER_NUM;
+const size_t PCMDATA_CACHEMAXSIZE = 2621440;
+const size_t QUEUEBUFFER_NUM = 4;
+const size_t CHUNK_SIZE_MAX = PCMDATA_CACHEMAXSIZE / QUEUEBUFFER_NUM;
 
 typedef std::vector<BYTE> PCMBuffer;
 
@@ -127,7 +128,7 @@ class MP3Reader : public AudioSourceReader
      HRESULT readAudioData(IMFSourceReader* pReader);
      void chunkify(PCMBuffer& buffer);
      bool appendToMappedWavFile(PCMBuffer& buffer);
-     void readFromMappedWavFile(BYTE *data, size_t offset, int size, UINT *pRetSize);
+     void readFromMappedWavFile(BYTE *data, size_t offset, size_t size, UINT *pRetSize);
      Microsoft::WRL::Wrappers::FileHandle openFile(const std::string& path, bool append = false);
 
  private:
